Django-Rest-Framework/requirements.txt
absl-py==0.7.1
amqp==2.4.2
asn1crypto==0.24.0
astor==0.7.1
Babel==2.6.0
bcrypt==3.1.6
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.2.1
djangorestframework==3.9.3
elasticsearch==7.0.0
flower==0.9.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
imutils==0.5.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kombu==4.5.0
Markdown==3.1
mock==3.0.4
mysqlclient==1.4.2.post1
nltk==3.4
numpy==1.16.3
opencv-python==4.0.0.21
pandas==0.24.2
pandasticsearch==0.5.1
paramiko==2.4.2
Pillow==5.4.1
protobuf==3.7.1
pyasn1==0.4.5
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==5.1.1
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.2
